chore: Move cleanup functions from `audio.ts`
