+++ mobile/src/modules/scanning/helpers/audio.ts
@@ -2,32 +2,19 @@ import {
   MetadataPresets,
   getMetadata,
 } from "@missingcore/react-native-metadata-retriever";
-import { inArray, lt } from "drizzle-orm";
+import { inArray } from "drizzle-orm";
 import { File } from "expo-file-system";
 import type { Asset as MediaLibraryAsset } from "expo-media-library";
 import { getAssetsAsync } from "expo-media-library";
 
 import { db } from "~/db";
 import type { InvalidTrack } from "~/db/schema";
-import {
-  albums,
-  albumsToArtists,
-  artists,
-  hiddenTracks,
-  invalidTracks,
-  playedMediaLists,
-  tracks,
-  tracksToArtists,
-  tracksToPlaylists,
-  waveformSamples,
-} from "~/db/schema";
+import { invalidTracks, tracks, tracksToArtists } from "~/db/schema";
 
 import { upsertAlbums } from "~/api/album";
 import { AlbumArtistsKey } from "~/api/album.utils";
 import { createArtists } from "~/api/artist";
-import { RECENT_RANGE_MS } from "~/api/recent";
 import { upsertTracks } from "~/api/track";
-import { Queue } from "~/stores/Playback/actions";
 import { preferenceStore } from "~/stores/Preference/store";
 import { onboardingStore } from "../services/Onboarding";
 
@@ -182,82 +169,6 @@ export async function findAndSaveAudio() {
     unstagedFiles: unstagedTracks,
     changed: stats.unstaged,
   };
-}
-//#endregion
-
-//#region Cleanup Functions
-/** Clean up all unused content from a validated list of found content. */
-export async function cleanupDatabase(usedTrackIds: string[]) {
-  // Remove any unused tracks.
-  const unusedTrackIds = (
-    await Promise.all(
-      [hiddenTracks, invalidTracks, tracks].map((sch) =>
-        db.select({ id: sch.id }).from(sch),
-      ),
-    )
-  )
-    .flatMap((ids) => ids.map(({ id }) => id))
-    .filter((id) => !usedTrackIds.includes(id));
-  if (unusedTrackIds.length > 0) {
-    await Promise.allSettled([
-      db.delete(hiddenTracks).where(inArray(hiddenTracks.id, unusedTrackIds)),
-      db.delete(invalidTracks).where(inArray(invalidTracks.id, unusedTrackIds)),
-      db.delete(tracks).where(inArray(tracks.id, unusedTrackIds)),
-      db
-        .delete(tracksToArtists)
-        .where(inArray(tracksToArtists.trackId, unusedTrackIds)),
-      db
-        .delete(tracksToPlaylists)
-        .where(inArray(tracksToPlaylists.trackId, unusedTrackIds)),
-      db
-        .delete(waveformSamples)
-        .where(inArray(waveformSamples.trackId, unusedTrackIds)),
-    ]);
-  }
-
-  // Clear the queue of deleted tracks.
-  await Queue.removeIds(unusedTrackIds);
-
-  // Remove anything else that's unused.
-  await removeUnusedCategories();
-}
-
-/** Remove any albums or artists that aren't used. */
-export async function removeUnusedCategories() {
-  // Remove recently played media that's beyond what we display.
-  await db
-    .delete(playedMediaLists)
-    .where(lt(playedMediaLists.lastPlayedAt, Date.now() - RECENT_RANGE_MS));
-
-  // Remove unused albums.
-  const allAlbums = await db.query.albums.findMany({
-    columns: { id: true },
-    with: { tracks: { columns: { id: true }, limit: 1 } },
-  });
-  const unusedAlbumIds = allAlbums
-    .filter(({ tracks }) => tracks.length === 0)
-    .map(({ id }) => id);
-  await db
-    .delete(albumsToArtists)
-    .where(inArray(albumsToArtists.albumId, unusedAlbumIds));
-  await db.delete(albums).where(inArray(albums.id, unusedAlbumIds));
-
-  // Remove unused artists.
-  const allArtists = await db.query.artists.findMany({
-    columns: { name: true },
-    with: {
-      //? Relations used to filter out artists with no albums & tracks.
-      albumsToArtists: { columns: { albumId: true }, limit: 1 },
-      tracksToArtists: { columns: { trackId: true }, limit: 1 },
-    },
-  });
-  const unusedArtistNames = allArtists
-    .filter(
-      ({ albumsToArtists, tracksToArtists }) =>
-        albumsToArtists.length === 0 && tracksToArtists.length === 0,
-    )
-    .map(({ name }) => name);
-  await db.delete(artists).where(inArray(artists.name, unusedArtistNames));
 }
 //#endregion
