+++ mobile/src/modules/scanning/helpers/cleanup.ts
@@ -1,8 +1,23 @@
-import { isNotNull } from "drizzle-orm";
+import { inArray, isNotNull, lt } from "drizzle-orm";
 import { Directory } from "expo-file-system";
 
 import { db } from "~/db";
-import { albums, artists, playlists, tracks } from "~/db/schema";
+import {
+  albums,
+  albumsToArtists,
+  artists,
+  hiddenTracks,
+  invalidTracks,
+  playedMediaLists,
+  playlists,
+  tracks,
+  tracksToArtists,
+  tracksToPlaylists,
+  waveformSamples,
+} from "~/db/schema";
+
+import { RECENT_RANGE_MS } from "~/api/recent";
+import { Queue } from "~/stores/Playback/actions";
 
 import { ImageDirectory, deleteImage } from "~/lib/file-system";
 import { batch } from "~/utils/promise";
@@ -49,3 +64,77 @@ export async function cleanupImages() {
   console.log(`Deleted ${deletedCount} unlinked images.`);
 }
 //#endregion
+
+/** Clean up all unused content from a validated list of found content. */
+export async function cleanupDatabase(usedTrackIds: string[]) {
+  // Remove any unused tracks.
+  const unusedTrackIds = (
+    await Promise.all(
+      [hiddenTracks, invalidTracks, tracks].map((sch) =>
+        db.select({ id: sch.id }).from(sch),
+      ),
+    )
+  )
+    .flatMap((ids) => ids.map(({ id }) => id))
+    .filter((id) => !usedTrackIds.includes(id));
+  if (unusedTrackIds.length > 0) {
+    await Promise.allSettled([
+      db.delete(hiddenTracks).where(inArray(hiddenTracks.id, unusedTrackIds)),
+      db.delete(invalidTracks).where(inArray(invalidTracks.id, unusedTrackIds)),
+      db.delete(tracks).where(inArray(tracks.id, unusedTrackIds)),
+      db
+        .delete(tracksToArtists)
+        .where(inArray(tracksToArtists.trackId, unusedTrackIds)),
+      db
+        .delete(tracksToPlaylists)
+        .where(inArray(tracksToPlaylists.trackId, unusedTrackIds)),
+      db
+        .delete(waveformSamples)
+        .where(inArray(waveformSamples.trackId, unusedTrackIds)),
+    ]);
+  }
+
+  // Clear the queue of deleted tracks.
+  await Queue.removeIds(unusedTrackIds);
+
+  // Remove anything else that's unused.
+  await removeUnusedCategories();
+}
+
+/** Remove any albums or artists that aren't used. */
+export async function removeUnusedCategories() {
+  // Remove recently played media that's beyond what we display.
+  await db
+    .delete(playedMediaLists)
+    .where(lt(playedMediaLists.lastPlayedAt, Date.now() - RECENT_RANGE_MS));
+
+  // Remove unused albums.
+  const allAlbums = await db.query.albums.findMany({
+    columns: { id: true },
+    with: { tracks: { columns: { id: true }, limit: 1 } },
+  });
+  const unusedAlbumIds = allAlbums
+    .filter(({ tracks }) => tracks.length === 0)
+    .map(({ id }) => id);
+  await db
+    .delete(albumsToArtists)
+    .where(inArray(albumsToArtists.albumId, unusedAlbumIds));
+  await db.delete(albums).where(inArray(albums.id, unusedAlbumIds));
+
+  // Remove unused artists.
+  const allArtists = await db.query.artists.findMany({
+    columns: { name: true },
+    with: {
+      //? Relations used to filter out artists with no albums & tracks.
+      albumsToArtists: { columns: { albumId: true }, limit: 1 },
+      tracksToArtists: { columns: { trackId: true }, limit: 1 },
+    },
+  });
+  const unusedArtistNames = allArtists
+    .filter(
+      ({ albumsToArtists, tracksToArtists }) =>
+        albumsToArtists.length === 0 && tracksToArtists.length === 0,
+    )
+    .map(({ name }) => name);
+  await db.delete(artists).where(inArray(artists.name, unusedArtistNames));
+}
