+++ mobile/src/navigation/screens/tracks/ModifyView.tsx
@@ -25,9 +25,11 @@ import { updateTrack } from "~/api/track";
 import { useTrack } from "~/queries/track";
 import { Resynchronize } from "~/stores/Playback/actions";
 import { usePreferenceStore } from "~/stores/Preference/store";
-import { removeUnusedCategories } from "~/modules/scanning/helpers/audio";
 import { getArtworkUri } from "~/modules/scanning/helpers/artwork";
-import { cleanupImages } from "~/modules/scanning/helpers/cleanup";
+import {
+  cleanupImages,
+  removeUnusedCategories,
+} from "~/modules/scanning/helpers/cleanup";
 import { FormStateProvider, useFormStateContext } from "~/hooks/useFormState";
 
 import { useFloatingContent } from "~/navigation/hooks/useFloatingContent";
